+++ includes/service-providers/constant_contact/class-newspack-newsletters-constant-contact-sdk.php
@@ -734,7 +734,12 @@ final class Newspack_Newsletters_Constant_Contact_SDK {
 		$body    = [];
 		if ( $contact && ! \is_wp_error( $contact ) ) {
 			$body = [
-				'email_address'    => get_object_vars( $contact->email_address ),
+				'email_address'    => isset( $data['email_address'] ) ?
+					[
+						'address'            => $data['email_address'],
+						'permission_to_send' => 'implicit',
+					] :
+					get_object_vars( $contact->email_address ),
 				'list_memberships' => $contact->list_memberships,
 				'custom_fields'    => array_map( 'get_object_vars', $contact->custom_fields ),
 				'update_source'    => 'Contact',
@@ -785,20 +790,6 @@ final class Newspack_Newsletters_Constant_Contact_SDK {
 			if ( isset( $data['taggings'] ) ) { // Using isset and not empty because this can be an empty array.
 				$body['taggings'] = $data['taggings'];
 			}
-			if ( isset( $body['update_source'], $data['email'] ) ) {
-				$body['email_address']['address'] = $data['email'];
-				$existing_contact                 = $this->get_contact( $data['email'] );
-				if ( $existing_contact && ! \is_wp_error( $existing_contact ) ) {
-					try {
-						// If we are updating to an existing email address, delete the old one and update the existing one.
-						$this->request( 'DELETE', 'contacts/' . $contact->contact_id );
-					} catch ( Exception $e ) {
-						Newspack_Newsletters_Logger::log( 'Error deleting contact during upsert with email ' . $data['email'] . ': ' . $e->getMessage() );
-					}
-				}
-				$contact       = $existing_contact;
-				$email_address = $data['email'];
-			}
 		}
 
 		try {
@@ -810,7 +801,7 @@ final class Newspack_Newsletters_Constant_Contact_SDK {
 		} catch ( Exception $e ) {
 			return new WP_Error( 'newspack_newsletters_constant_contact_api_error', $e->getMessage() );
 		}
-		return $this->get_contact( $email_address );
+		return $res;
 	}
 
 	/**