+++ includes/service-providers/constant_contact/class-newspack-newsletters-constant-contact.php
@@ -1205,8 +1205,8 @@ final class Newspack_Newsletters_Constant_Contact extends \Newspack_Newsletters_
 		if ( isset( $contact['existing_contact_data']['email_address'] ) ) {
 			$existing_email = $contact['existing_contact_data']['email_address'];
 			if ( $existing_email->address !== $email ) {
-				$data['email'] = $email;
-				$email         = $existing_email->address;
+				$data['email_address'] = $email;
+				$email                 = $existing_email->address;
 			}
 		}
 		$result = $cc->upsert_contact( $email, $data );